Do not read the entry to compute size #181

+++ core/org.eclipse.rcptt.core/src/org/eclipse/rcptt/core/persistence/IPersistenceModel.java
@@ -95,7 +95,7 @@ public interface IPersistenceModel {
 
 	Resource getResource();
 
-	int size(String teslaContentEntry);
+	long size(String teslaContentEntry);
 
 	boolean isContentEntryRequired();
